Add Sound FX for the RMC coasters
These sound effects are intended to be used for the Hybrid RC and the Single Rail RC to differentiate in the base game (so they should be available without asset pack loaded).

For these two coaster types only 56 would be replaced with the lift and 57 would be replaced with the roar.

If you wish to know how it sounds like you can check it out in discord: https://discord.com/channels/264137540670324737/694586451500859469/1022699064841748532