Add airbrake sound
This is the hydrolic sound of a towtruck. It's the same mechanism that is used to pressure the friction brakes in real life. I have not added this in the json yet as it's too monotone at this moment in time. Might look into re-recording this from actual friction brakes and have more different sounds.