Add FX for B&M, Toilets and Airbrakes
* Add replacement sounds for B&M

These sounds are both from rougarou. I did hear a complain someone didn't like the ticking as much, but it's accurate with this specific coaster. Could always look for a new B&M lift in the future.

* Add toilet flush sound

This is a flush sound for the bathrooms. This specific sound is a recording of an American home toilet. I think the original sound is from a public pressure toilet in the UK or France. (Due to the whitenoise). I think it sounds quite well with the other sounds.

* Add airbrake sound

This is the hydrolic sound of a towtruck. It's the same mechanism that is used to pressure the friction brakes in real life. I have not added this in the json yet as it's too monotone at this moment in time. Might look into re-recording this from actual friction brakes and have more different sounds.

* Update JSON

Fix formatting, somehow used Tab instead of Spaces.